Use `f32` instead of `f64`

+++ examples/cli.rs
@@ -10,8 +10,8 @@ use aeroscore::olc;
 
 struct Point {
     time: igc::util::Time,
-    latitude: f64,
-    longitude: f64,
+    latitude: f32,
+    longitude: f32,
     altitude: i16,
 }
 
@@ -35,10 +35,10 @@ impl std::fmt::Debug for Point {
 }
 
 impl aeroscore::Point for Point {
-    fn latitude(&self) -> f64 {
+    fn latitude(&self) -> f32 {
         self.latitude
     }
-    fn longitude(&self) -> f64 {
+    fn longitude(&self) -> f32 {
         self.longitude
     }
     fn altitude(&self) -> i16 {
